Move the scripts to a separate file in "includes" folder
